+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## [NEXT] - YYYY-MM-DD
-- Add command to dump item stack render type data (currently not usable)
+- Add command to dump item stack render type data
 - Adjust render type heuristic: Signs, beds and shields now use item rendering
 - Fix tile rotation not being read
 - Adjust rendering to no longer change lighting state